Move expandTilda to pkg/helpers/path as exported ExpandTilda
Extract path utility from pkg/config/util.go into a dedicated
pkg/helpers/path package, reducing config's role as a utility
dumping ground. Update all call sites to use pathutil.ExpandTilda.
